requirements.txt
pandas==2.1.4
lxml==4.9.3
azure-storage-blob==12.19.0